pyproject.toml
[tool.poetry]
name = "ansible-docs"
version = "0.1.0"
description = "A tool for generating docs for Ansible roles"
authors = ["Miika Kankare <miika@kanka.re>"]
license = "MIT"
packages = [
  { include = "ansible_docs" }
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.950"
pytest = "^7.1.3"

[tool.poetry.scripts]
ansible-docs = "ansible_docs.cli:app"

[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib",
]
pythonpath = "."


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
